Fix with-contract when used in combination with keyword contracts/definitions.
This change points out with-contract's current brittleness re: define
macros.  I'm not sure how to best fix this, though, without returning
to either the thunk version of with-contract (which has its own
problems, namely that the external id is defined well after the
internal one) or without having something like a guaranteed form I can
look for that says "this block of code is an expansion for defining
this identifier", so I can place the contract/external id defines
immediately after that block.